Added sync, save and load
Now the API will track the state of pixels, allowing clients to re-sync their view to the server state. This allows the session to survive refreshes, and resuming drawing from another device.

Also added facility to save state to a server-side file, which the client can request. The API will re-draw the the last saved state on startup, making drawings persistent until the next save.